pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trestle-mcp"
version = "0.1.0"
description = "MCP server for compliance-trestle OSCAL framework"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "Trestle MCP Contributors"}
]
keywords = ["mcp", "claude", "oscal", "trestle", "compliance", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "compliance-trestle>=3.11.0",
    "ruamel.yaml>=0.19.0"
]

[project.optional-dependencies]
dev = [
    "build",
    "deepdiff==8.1.1",
    "fastmcp==2.14.4",
    "pyclean",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-timeout==2.3.1",
    "pre-commit>=4.2.0",
    "black",
    "isort",
    "tomli-w",
]

[project.urls]
Homepage = "https://github.com/yourusername/trestle-mcp"
Repository = "https://github.com/yourusername/trestle-mcp.git"
Issues = "https://github.com/yourusername/trestle-mcp/issues"

[project.scripts]
trestle-mcp = "trestle_mcp.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]


[tool.black]
target-version = [
    "py39",
    "py310",
    "py311",
]
force-exclude = "/(\n\\.venv\n)/\n"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.ruff]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "tests",
]

[tool.ruff.lint]
ignore = [
    "E722",
    "F401",
]